fix(agents): make a short-lived certificate something a deployment can ask for
Agents renew at two-thirds of certificate life. AGENT_CERT_TTL_HOURS was an
integer, so the shortest certificate a *deployment* could issue was an hour and
its renewal point was forty minutes away — beyond the reach of any harness that
stands the platform up and watches it. The property was covered only by tests
that construct AgentIdentityService(leaf_lifetime=timedelta(seconds=30))
directly, which proves the mechanism but not that a deployment can be
configured into it.

The setting is now a float. AGENT_CERT_TTL_HOURS=0.025 is ninety seconds,
renewing at sixty; existing integer values parse unchanged and the default is
still 90 days.

Asserted on the validity of a certificate build_identity_service() actually
issues, not on the setting — a setting that parses and is never passed on reads
identically to a working one, which is the shape the X-Scope-OrgID header
already had here. Mutation-tested by reverting the annotation to int: caught.

This does not add the CI check; it removes what blocked one.
docs/PRODUCTION_READINESS.md now says so, and corrects the earlier claim that
renewal is uncovered — it is covered and passing, by a Go test against an
arbitrary clock and by a real-binary test against a real gateway.

+++ backend/app/core/config.py
@@ -433,7 +433,16 @@ class Settings(BaseSettings):
 
     # Issued certificate life, per ADR-005. Agents renew at 2/3 of it, so the
     # default leaves a 30-day overlap window.
-    agent_cert_ttl_hours: int = Field(default=24 * 90, validation_alias="AGENT_CERT_TTL_HOURS")
+    #
+    # A **float**, and the reason is testability rather than anyone wanting
+    # fractional hours in production. As an `int` the shortest deployable
+    # certificate was one hour, whose renewal point is forty minutes away —
+    # which put renewal out of reach of any harness that stands the platform up
+    # and watches it, and left the property covered only by tests that
+    # construct `AgentIdentityService` directly. `AGENT_CERT_TTL_HOURS=0.025`
+    # is ninety seconds, renewing at sixty. Existing integer values parse
+    # unchanged.
+    agent_cert_ttl_hours: float = Field(default=24 * 90, validation_alias="AGENT_CERT_TTL_HOURS")
 
     # How often the gateway re-reads the revocation list and drops any live
     # stream whose certificate has since been revoked. Revocation that only
